Refactor to one Component per file
