+++ src/components/about/index.ts
@@ -1,0 +1,5 @@
+import { AboutMe } from './AboutMe'
+
+export {
+    AboutMe
+}
