+++ src/components/contact/index.ts
@@ -1,0 +1,5 @@
+import { ContactMe } from './ContactMe'
+
+export {
+    ContactMe
+}
