+++ src/components/workingOn/index.ts
@@ -1,0 +1,5 @@
+import { WorkingOn } from "./WorkingOn";
+
+export {
+    WorkingOn
+}
